added celery start file

+++ fabfile.py
@@ -144,9 +144,10 @@ def _setup_nginx_server():
             env.host
         ))
 
-    sudo('ln /etc/nginx/sites-available/{}.conf /etc/nginx/sites-enabled/'.format(
-        env.host
-    ))
+    if not exists('/etc/nginx/sites-enabled/{}.conf'):
+        sudo('ln /etc/nginx/sites-available/{}.conf /etc/nginx/sites-enabled/'.format(
+            env.host
+        ))
     sudo('service nginx restart')
 
 
@@ -160,6 +161,7 @@ def _setup_supervisor():
     sudo('supervisorctl update')
     sudo('supervisorctl start {}'.format(env.host))
     sudo('supervisorctl start celery')
+
 
 def _create_celery_directories():
     sudo('mkdir -p /var/log/celery')
